Isolated forwardtoolstep() and it's supporting methods
Isolated forwardtoolstep() and it's supporting methods
forwardtoolstep() reads from the last computed state now and should be fed any inputs it needs. The first state needs to be initialized with createmotionfilestep() outside of forwardtoolstep() though.
Changed the file paths by replacing Project with controlled-movement
Added a .gitignore
Generates a bicep curl (0.03 to 0.489) in 15.77 seconds, down from 20.26