Huge Update
Added MonkeyArm files to try and test a bicep curl, the .osim file may have problems - Noman says it might be the muscle definition
Tried to make an s-function, may be removed in a future version.
Tried my hand at creating a better simulink environment, simulink most likely will not work.
Created a loop version of the forwardtool which returns the state of the coordinates
Created 'better' versions of the forwardtool supporting methods
Struggled to understand continuous PID, created an algorithm of the discrete form thanks to Wikipedia - PID_controller#Pseudocode
Moved random code to a separate folder, another cleanup is still needed, will do that in a future update