Function Generator
Using str2func, I devised a way to convert piecewise functions in vector notation into a single function handle
logicalindexingtest has some interesting attempts at this and various functions I created defined in this file
str2functest is the beginning of a system that can take the raw parameters of a piecewise function and convert it into a string, convert it into an anonymous function handle, store it, and then use it. Mwah!