Function Reader Works
This update enables functions to be easily defined in a file in piecewise component form and then translated by MATLAB into function handles!
I created functions.txt to define the functions and added a guide to the file to understand the format and how to use it.
I created a piecewise function object class called pwfunc to temporarily store the arguments and bounds for each function and to make the code look cleaner and more readable.
The function to read the functions.txt file and create pwfunc objects took a while to make, but it now works probably flawlessly.
I rewrote str2functest (name change pending) to turn it into a for loop to iterate through each function object and turn it into a quick object that stores a function handle, the start time, and end time.
This new object will become what I work with and make the kinematics files off of.
TODO: Rewrite the SIdatagen file to use this new function handle system.