SI Data Gen Update
Patch Notes:
Changed the name of motionheaderwriter to motionfile_writer.
Changed the name of str2functest to motion_properties_generator.
Need to add function descriptions for the latest files.
SIdatagen breaks at function 6 as the function is not made properly and is too much for the biomechanics of the arm. I need to rework function 6.
Reworked SIdatagen to use the new motion properties struct, which simplified the workflow.
Created motion file plotter to plot the positions and velocity of the joint movements for debugging use.
Enabled setting of initial and final time for RRA, CMC, and FD tool to allow for time ranges that don’t match the setup files.