FD Broken, Going with CMC
The Forward Tool isn't providing correct results and instead is providing results that appear to have memory of past results, consult the figures FD Graphs 1 and 2 for the weirdness.
Created the function raw_vs_cmc_error to visualize the error between the desired results and the CMC and FD results. They're very similar so for the first run I'm going to use the CMC states as the FD states and run system identification on that.
Changed SIdatagen to allow for the tools to break and not cause the script to stop but give a soft error instead.
To support this behavior, I changed the RRA, CMC, and FD tool to return a success value.
I fixed two offending functions in functions.txt. I totally miscalculated the sinusoids which caused OpenSim to break since it was not biomechanically possible movement.
I created states_error_grapher to graph the states between all files to visually observe the difference. I created the runforall method as a way to run it for each file at once to compare in bulk. This is what showcased the weirdness of the FD tool. FD Graphs 1 and 2 are the output of this function.