Auto Motion Creator
Created a system to automagically generate features within an elbow movement for a certain time period. Really helps with bulk creation of test data.
Took a couple files and tests to work my way towards the final version which is automotioncreator.m
More info on next update, need to see if this build works and I want to use this as a save point.