Auto Motion Creator Works :tm:
Cleaned up System Identification file directories. Moved them off of the cloud to speed up processing and space. Their respective data files will be maintained in the cloud, but the bulk of thousands of files will not be.

The auto motion creator works by creating features in a specific time interval, adding constant lines, linear slopes, or parabolas to fill the various features, and then smoothing it out with a moving average filter with a span of 40.

TODO: Add support for testing and validation files, probably as a change to the file's names.

- Other Changes -
Removed a lot of junk like FD Graphs, osimloop, test scripts, the old motion files still in the main dir, and all the sys id bulk files.
Created a grapher to show all the auto created functions and their respective CMC states functions.
Changed file paths to redirect to my SSD where the bulk files will be stored.