HIPERGATOR
WITH THE COMPUTATIONAL POWER OF HIPERGATOR, WE SHALL BE KINGS.

All craziness aside, holy crap HiPerGator is fast. Took a long time to figure out how to do arrays, assignment, ceil, floor, mod, echo, and other operations in bash. But I figured it out and now I can run job array scripts which saves my fingers from having to type!

Uploaded all my HiPerGator Matlab stuff (HPGM folder). This contains the sbatch scripts for slurm as well as other things. This folder is for HiPerGator only and is synced back and forth with Globus file manager on the web.

Also I moved all old scripts to the new Old Scripts folder so there would be less clutter. I can pull them out if needed or maybe they'll stay in that archive for a while.

Added opensim.log to git ignore so it doesn't keep updating!

Created a new comparesystems script to compare the various orders of the systems. But I'm going to have to expand on this to support the crazy new amounts of numbers the systems use.

I had to ignore all .mat files as they were > 100 MB and thus GitHub didn't like that and I don't feel like figuring out GitHub LFS tonight.