HPGM Pipeline
Got a pipeline approach with job dependencies so most of the work is automated. Can't figure out being able to add sidp3 as it requires knowing if the child jobs of tpipe and vpipe have finished. Can't find any literature on if that's possible and if so, how.

Created the 500, 1000, 2500, and 5000 experiments for 2, 4, 6, 8, and 10 orders. Order 10 was proving to be the best and now I'm moving on to examine orders 10 through 15 to see if there's any improvement.

tpipe handles all the job calls for the testing data and system generation

vpipe handles all the job calls for the validation data and system generation

Changed mcp1-3 and sidp1-3 to reflect the pipelining as I can now pass variables in. I removed any hardcoded elements within these job scripts to allow them to be flexible for both testing and validation data.