Live Graphs
Live graphing finally works.

In the previous update I had started a MATLAB App to control the reference signals for position and velocity, and visualize the output and reference signals.

Next goal is to use a MATLAB App to control the reference signals and visualize the results. I think I might use the following resource to help as it seems pretty simple: https://www.mathworks.com/matlabcentral/answers/441430-how-to-return-values-from-mlapp-to-the-m-caller