IT WORKS! IT ACTUALLY WORKS!
For 10th order system with frequency crossover range of [50, 500], I get a pretty blazing fast step response. Rise time of 0.009 seconds with a settling time of 0.047 seconds and an overshoot of 18.73%. On top of that, with the gain settings of Kp = 0.017 and Ki = 0.03, it becomes controllable. There's unsteady behavior in the shoulder joint and the arm oscillates a lot, but the Arm26 does a bicep curl and DOES NOT go into the negative range of motion which usually breaks it quickly.

Not the best results, I'd love something that could pump some iron. But results nonetheless! I can now write a convincing paper and draw some higher level conclusions.

Added variable control to top of both control_loop and controller_tuning to minimize the need for scrolling through hundreds of lines just to change the same variable.

Added ability for derivative control. Just takes the derivative of the last few n specified points.

Tried to make steptester function find the BEST stepfunction as it iterates through nearly every combination of wc values. Lot of stuff to fix here, mostly computational ability. Need to run it on HiPerGator honestly.

This documentation was a bit of a ramble, but hopefully it's better than Microsoft's!